producer/requirements.txt
azure-eventhub==5.11.0
